Remove non-official warning and add PR template
Climesync is now maintained by an OSL developer and will likely
be the primary tool for communicating with TimeSync for a while.
This should be reflected in the README.

Since this is becoming a useful too, we should us a Github PR
template to make testing easier. This is specifically important
for Climesync because it has one main developer. Reviewers will
not be familiar with the codebase and the clear outline the template
provides will be very helpful.